refactor: move inline tests to command_line_test.rs
Inline test modules in source files conflict with the project
convention of using separate _test.rs files. Move the Cli socket
tests to src/app/cli/command_line_test.rs (declared from mod.rs)
and widen resolved_socket() to pub(crate) so the sibling test
module can access it.
