+++ src/app/cli/mod.rs
@@ -3,3 +3,6 @@ mod completion;
 mod config;
 mod session;
 pub use command_line::Cli;
+
+#[cfg(test)]
+mod command_line_test;
